requirements.txt
hyperliquid_monitor==0.1.2.1
python-dotenv==1.1.0
Requests==2.32.3
